Queue messages sent before websocket was open.
Nothing stopped you from sending a message before the handshake is
complete and the websocket is ready. Before this change, doing so
could cause a crash if you were unlucky.

I have a codebase that relies on code like this to work:

```objective-c
PSWebSocket *socket = [PSWebSocket clientSocketWithRequest:request];
[socket open];
[socket send:@"test];
```

(I was using SwiftWebSocket before.)

I added a queue of messages sent before the handshake is complete,
which makes that safe.